Move RouterService into Services\ namespace
