+++ packages/realtime-compiler/src/Routing/Router.php
@@ -47,7 +47,7 @@ class Router
         }
 
         // Temporary backwards compatibility for versions less than Hyde/Framework v0.48.0-beta
-        if (! class_exists('\Hyde\Framework\RoutingService')) {
+        if (! class_exists('\Hyde\Framework\Services\RoutingService')) {
             return LegacyPageRouter::handle($this->request);
         }
